# Conflicts:
#	.github/workflows/prod.yml
#	.github/workflows/staging.yml
#	.gitignore
#	apps/api-runtime/src/main/java/com/akto/dependency/DependencyAnalyser.java
#	apps/api-runtime/src/main/java/com/akto/parsers/HttpCallParser.java
#	apps/api-runtime/src/main/java/com/akto/runtime/APICatalogSync.java
#	apps/api-runtime/src/main/java/com/akto/runtime/Main.java
#	apps/api-runtime/src/test/java/com/akto/parsers/TestDBSync.java
#	apps/api-runtime/src/test/java/com/akto/parsers/TestMergingNew.java
#	apps/dashboard/src/main/java/com/akto/action/CodeAnalysisAction.java
#	apps/dashboard/src/main/java/com/akto/action/observe/InventoryAction.java
#	apps/dashboard/src/main/java/com/akto/action/testing/StartTestAction.java
#	apps/dashboard/src/main/java/com/akto/listener/InitializerListener.java
#	apps/dashboard/src/main/java/com/akto/utils/Utils.java
#	apps/dashboard/web/polaris_web/web/src/apps/dashboard/pages/observe/api.js
#	apps/dashboard/web/polaris_web/web/src/apps/dashboard/pages/observe/api_collections/RunTest.jsx
#	apps/database-abstractor/src/main/java/com/akto/action/CodeAnalysisAction.java
#	apps/database-abstractor/src/main/java/com/akto/action/DbAction.java
#	apps/database-abstractor/src/main/java/com/akto/action/InfraMetricsAction.java
#	apps/mini-runtime/src/main/java/com/akto/hybrid_dependency/DependencyAnalyser.java
#	apps/mini-runtime/src/main/java/com/akto/hybrid_parsers/HttpCallParser.java
#	apps/mini-runtime/src/main/java/com/akto/hybrid_runtime/APICatalogSync.java
#	apps/mini-runtime/src/main/java/com/akto/hybrid_runtime/Main.java
#	apps/mini-runtime/src/main/java/com/akto/hybrid_runtime/policies/AktoPolicyNew.java
#	libs/dao/src/main/java/com/akto/DaoInit.java
#	libs/dao/src/main/java/com/akto/dao/filter/MergedUrlsDao.java
#	libs/dao/src/main/java/com/akto/dao/monitoring/FilterConfigYamlParser.java
#	libs/dao/src/main/java/com/akto/dao/monitoring/FilterYamlTemplateDao.java
#	libs/dao/src/main/java/com/akto/dao/traffic_metrics/RuntimeMetricsDao.java
#	libs/dao/src/main/java/com/akto/dto/AccountSettings.java
#	libs/dao/src/main/java/com/akto/dto/ApiCollection.java
#	libs/dao/src/main/java/com/akto/dto/CodeAnalysisApi.java
#	libs/dao/src/main/java/com/akto/dto/CodeAnalysisRepo.java
#	libs/dao/src/main/java/com/akto/dto/HttpResponseParams.java
#	libs/dao/src/main/java/com/akto/dto/filter/MergedUrls.java
#	libs/dao/src/main/java/com/akto/dto/monitoring/FilterConfig.java
#	libs/dao/src/main/java/com/akto/dto/traffic/SuspectSampleData.java
#	libs/dao/src/main/java/com/akto/dto/type/KeyTypes.java
#	libs/dao/src/main/java/com/akto/dto/type/URLTemplate.java
#	libs/dao/src/main/java/com/akto/util/filter/DictionaryFilter.java
#	libs/dao/src/main/java/com/akto/util/filter/DictionaryFilterCreator.java
#	libs/dao/src/test/java/com/akto/utils/filter/TestDictionaryFilter.java
#	libs/utils/src/main/java/com/akto/data_actor/ClientActor.java
#	libs/utils/src/main/java/com/akto/data_actor/DataActor.java
#	libs/utils/src/main/java/com/akto/data_actor/DbActor.java
#	libs/utils/src/main/java/com/akto/data_actor/DbLayer.java
#	libs/utils/src/main/java/com/akto/log/LoggerMaker.java
#	libs/utils/src/main/java/com/akto/metrics/AllMetrics.java
#	libs/utils/src/main/java/com/akto/runtime/policies/ApiAccessTypePolicy.java
#	libs/utils/src/main/java/com/akto/testing/ApiExecutor.java
#	libs/utils/src/main/java/com/akto/usage/UsageMetricCalculator.java
#	libs/utils/src/main/java/com/akto/utils/CustomAuthUtil.java

+++ libs/dao/src/main/java/com/akto/dto/ApiInfo.java
@@ -338,8 +338,8 @@ public class ApiInfo {
 
     public void setId(ApiInfoKey id) {
         this.collectionIds = Util.replaceElementInList(this.collectionIds, 
-        id == null ? null : id.getApiCollectionId(),
-        this.id == null ? null : this.id.getApiCollectionId());
+        id == null ? null : (Integer)id.getApiCollectionId(),
+        this.id == null ? null : (Integer)this.id.getApiCollectionId());
         this.id = id;
     }
